websocket and reservationService and sendEmail

+++ src/main/java/com/example/projectspring/entity/Reservation.java
@@ -8,7 +8,7 @@ public class Reservation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "trajet_id", nullable = false)
@@ -27,13 +27,11 @@ public class Reservation {
 
     @Column
     private String annulationJustification;
-    // Constructeurs
 
-    // Constructeur sans arguments (nécessaire pour JPA)
+    // Constructeurs
     public Reservation() {
     }
 
-    // Constructeur avec tous les arguments
     public Reservation(Trajet trajet, Users passager, LocalDateTime dateReservation, StatusReservation status) {
         this.trajet = trajet;
         this.passager = passager;
@@ -42,11 +40,11 @@ public class Reservation {
     }
 
     // Getters and Setters
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -81,6 +79,7 @@ public class Reservation {
     public void setStatus(StatusReservation status) {
         this.status = status;
     }
+
     public String getAnnulationJustification() {
         return annulationJustification;
     }
